fix: add `cpm_check_converted` to help fix `get_cpm_for_coord_alignment`
